Move EncyptedGroupKey to message_layer package
Fixes last cyclic dependency.
