+++ client/src/main/java/com/streamr/client/protocol/message_layer/EncryptedGroupKey.java
@@ -1,8 +1,7 @@
-package com.streamr.client.stream;
+package com.streamr.client.protocol.message_layer;
 
-import com.streamr.client.java.util.Objects;
-import com.streamr.client.protocol.message_layer.EncryptedGroupKeyAdapter;
 import java.io.IOException;
+import java.util.Objects;
 
 /**
  * A container for encrypted group keys. Used to get compile-time checking that encrypted and
