+++ client/src/main/java/com/streamr/client/stream/MessageCreationUtil.java
@@ -14,13 +14,7 @@ import com.streamr.client.protocol.message_layer.MessageId;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.protocol.message_layer.StringOrMillisDateJsonAdapter;
 import com.streamr.client.rest.Stream;
-import com.streamr.client.stream.EncryptedGroupKey;
-import com.streamr.client.stream.EncryptionUtil;
-import com.streamr.client.stream.GroupKey;
-import com.streamr.client.stream.InvalidGroupKeyException;
-import com.streamr.client.stream.InvalidGroupKeyRequestException;
-import com.streamr.client.stream.InvalidGroupKeyResponseException;
-import com.streamr.client.stream.KeyExchangeUtil;
+import com.streamr.client.protocol.message_layer.EncryptedGroupKey;
 import com.streamr.client.utils.SigningRequiredException;
 import com.streamr.client.utils.SigningUtil;
 import com.streamr.ethereum.common.Address;
