+++ client/src/test/groovy/com/streamr/client/protocol/message_layer/GroupKeyAnnounceAdapterSpec.groovy
@@ -1,6 +1,6 @@
 package com.streamr.client.protocol.message_layer
 
-import com.streamr.client.stream.EncryptedGroupKey
+
 import spock.lang.Specification
 
 class GroupKeyAnnounceAdapterSpec extends Specification {
