+++ client/src/test/java/com/streamr/client/protocol/message_layer/EncryptedGroupKeyTest.java
@@ -1,4 +1,4 @@
-package com.streamr.client.stream;
+package com.streamr.client.protocol.message_layer;
 
 import nl.jqno.equalsverifier.EqualsVerifier;
 import org.junit.jupiter.api.Test;
